# simd image blend when possible