… tests/images/rotate0.png
… tests/images/rotate180.png
… tests/images/rotate270.png
… tests/images/rotate360.png
… tests/images/rotate90.png