before (`nim c -d:pixieNoSimd -d:release -r .\tests\benchmark_images_draw.nim`)
```
name ............................... min time      avg time    std dv  times
draw small-on-big bmNormal ......... 0.021 ms      0.021 ms    ±0.001  x1000 
draw big-on-bigger bmNormal ........ 1.761 ms      1.827 ms    ±0.122  x1000 
```
after (`nim c -d:release -r .\tests\benchmark_images_draw.nim`)
```
name ............................... min time      avg time    std dv  times  
draw small-on-big bmNormal ......... 0.008 ms      0.008 ms    ±0.000  x1000
draw big-on-bigger bmNormal ........ 0.379 ms      0.389 ms    ±0.021  x1000
```